+++ src/main/java/io/github/sakurawald/core/config/model/ConfigModel.java
@@ -241,6 +241,7 @@ public class ConfigModel {
             public SendChat send_chat = new SendChat();
             public SendBossbar send_bossbar = new SendBossbar();
             public SendCustom send_custom = new SendCustom();
+            public SendConsole send_console = new SendConsole();
 
             public static class SendMessage {
                 public boolean enable = true;
@@ -271,6 +272,10 @@ public class ConfigModel {
             }
 
             public static class SendCustom {
+                public boolean enable = true;
+            }
+
+            public static class SendConsole {
                 public boolean enable = true;
             }
         }